+++ traffic_monitor/config/config_test.go
@@ -33,8 +33,6 @@ const exampleTMConfig = `
 	"peer_optimistic": true,
 	"peer_optimistic_quorum_min": 0,
 	"max_events": 200,
-	"max_stat_history": 5,
-	"max_health_history": 5,
 	"health_flush_interval_ms": 1000,
 	"stat_flush_interval_ms": 1000,
 	"log_location_event": "event.log",
@@ -45,7 +43,6 @@ const exampleTMConfig = `
 	"serve_read_timeout_ms": 10000,
 	"serve_write_timeout_ms": 10000,
 	"stat_buffer_interval_ms": 20000,
-	"http_poll_no_sleep": false,
 	"static_file_dir": "static/"
 }
 `